Application of eX mods

+++ Vault/VaultDataModels/Models/ITransaction.cs
@@ -8,7 +8,7 @@
 
         string Description { get; }
 
-        double Amount { get; }
+        decimal Amount { get; }
 
         DateTime ExecutionDate { get; }
     }
